refactor(tests): clean-up and group Earned Leave tests
