+++ hrms/hr/doctype/leave_application/test_leave_application.py
@@ -710,27 +710,6 @@ class TestLeaveApplication(unittest.TestCase):
 		self.assertEqual(details.leave_balance_for_consumption, 21)
 		self.assertEqual(details.leave_balance, 30)
 
-	def test_earned_leaves_creation(self):
-		from hrms.hr.utils import allocate_earned_leaves
-
-		leave_period = get_leave_period()
-		employee = get_employee()
-		leave_type = "Test Earned Leave Type"
-		make_policy_assignment(employee, leave_type, leave_period)
-
-		for i in range(0, 14):
-			allocate_earned_leaves()
-
-		self.assertEqual(get_leave_balance_on(employee.name, leave_type, nowdate()), 6)
-
-		# validate earned leaves creation without maximum leaves
-		frappe.db.set_value("Leave Type", leave_type, "max_leaves_allowed", 0)
-
-		for i in range(0, 6):
-			allocate_earned_leaves()
-
-		self.assertEqual(get_leave_balance_on(employee.name, leave_type, nowdate()), 9)
-
 	# test to not consider current leave in leave balance while submitting
 	def test_current_leave_on_submit(self):
 		employee = get_employee()
@@ -945,54 +924,6 @@ class TestLeaveApplication(unittest.TestCase):
 		self.assertEqual(leave_allocation["remaining_leaves"], 26)
 
 	@set_holiday_list("Salary Slip Test Holiday List", "_Test Company")
-	def test_get_earned_leave_details_for_dashboard(self):
-		from hrms.hr.utils import allocate_earned_leaves
-
-		leave_period = get_leave_period()
-		employee = get_employee()
-		leave_type = "Test Earned Leave Type"
-		leave_policy_assignments = make_policy_assignment(employee, leave_type, leave_period)
-		allocation = frappe.db.get_value(
-			"Leave Allocation",
-			{"leave_policy_assignment": leave_policy_assignments[0]},
-			"name",
-		)
-		allocation = frappe.get_doc("Leave Allocation", allocation)
-		allocation.new_leaves_allocated = 2
-		allocation.save()
-
-		for i in range(0, 6):
-			allocate_earned_leaves()
-
-		first_sunday = get_first_sunday(self.holiday_list)
-		make_leave_application(
-			employee.name, add_days(first_sunday, 1), add_days(first_sunday, 1), leave_type
-		)
-
-		details = get_leave_details(employee.name, allocation.from_date)
-		leave_allocation = details["leave_allocation"][leave_type]
-		expected = {
-			"total_leaves": 2.0,
-			"expired_leaves": 0.0,
-			"leaves_taken": 1.0,
-			"leaves_pending_approval": 0.0,
-			"remaining_leaves": 1.0,
-		}
-		self.assertEqual(leave_allocation, expected)
-
-		details = get_leave_details(employee.name, getdate())
-		leave_allocation = details["leave_allocation"][leave_type]
-
-		expected = {
-			"total_leaves": 5.0,
-			"expired_leaves": 0.0,
-			"leaves_taken": 1.0,
-			"leaves_pending_approval": 0.0,
-			"remaining_leaves": 4.0,
-		}
-		self.assertEqual(leave_allocation, expected)
-
-	@set_holiday_list("Salary Slip Test Holiday List", "_Test Company")
 	def test_get_leave_allocation_records(self):
 		employee = get_employee()
 		leave_type = create_leave_type(
@@ -1105,37 +1036,3 @@ def allocate_leaves(employee, leave_period, leave_type, new_leaves_allocated, el
 	).insert()
 
 	allocate_leave.submit()
-
-
-def make_policy_assignment(employee, leave_type, leave_period):
-	frappe.delete_doc_if_exists("Leave Type", leave_type, force=1)
-	frappe.get_doc(
-		dict(
-			leave_type_name=leave_type,
-			doctype="Leave Type",
-			is_earned_leave=1,
-			allocate_on="Last Day",
-			earned_leave_frequency="Monthly",
-			rounding=0.5,
-			max_leaves_allowed=6,
-		)
-	).insert()
-
-	leave_policy = frappe.get_doc(
-		{
-			"doctype": "Leave Policy",
-			"title": "Test Leave Policy",
-			"leave_policy_details": [{"leave_type": leave_type, "annual_allocation": 6}],
-		}
-	).insert()
-
-	data = {
-		"assignment_based_on": "Leave Period",
-		"leave_policy": leave_policy.name,
-		"leave_period": leave_period.name,
-	}
-
-	leave_policy_assignments = create_assignment_for_multiple_employees(
-		[employee.name], frappe._dict(data)
-	)
-	return leave_policy_assignments
